src/server/requirements_server.txt
numpy
pyserial
pycuda